actualiazacion clases x.10
cambio variable en terran

+++ src/juegorol/Terrans/BattleCruiser.java
@@ -14,7 +14,7 @@ public class BattleCruiser extends Terran{
          this.puntosunidad=4;
    }
    public BattleCruiser(float size,float cargo,int armor,int health,int ataque,
-                   int cooldown,int rango,int vista,int bonus,int puntosunidad)
+                   int cooldown,int rango,int vista,double bonus,int puntosunidad)
    {      
    }
 }
