+++ src/juegorol/Terrans/Marauder.java
@@ -14,7 +14,7 @@ public class Marauder extends Terran{
          this.puntosunidad=2;
    }
    public Marauder(float size,float cargo,int armor,int health,int ataque,
-                   int cooldown,int rango,int vista,int bonus,int puntosunidad)
+                   int cooldown,int rango,int vista,double bonus,int puntosunidad)
    {      
    }
 }
