+++ src/juegorol/Terrans/Reaper.java
@@ -14,7 +14,7 @@ public class Reaper extends Terran{
          this.puntosunidad=1;
    }
    public Reaper(float size,float cargo,int armor,int health,int ataque,
-                   int cooldown,int rango,int vista,int bonus,int puntosunidad)
+                   int cooldown,int rango,int vista,double bonus,int puntosunidad)
    {      
    }    
 }
